open_adr_3/requirements.txt
fastapi
uvicorn[standard]
PyJWT
python-multipart
